refactor: dto 디렉토리 구조 변경, 유효성 검사 메세지 통일 #25
