+++ src/main/java/com/canvi/hama/domain/auth/controller/EmailAuthController.java
@@ -3,8 +3,8 @@ package com.canvi.hama.domain.auth.controller;
 import com.canvi.hama.common.exception.BaseException;
 import com.canvi.hama.common.response.BaseResponse;
 import com.canvi.hama.common.response.BaseResponseStatus;
-import com.canvi.hama.domain.auth.dto.SendCodeRequest;
-import com.canvi.hama.domain.auth.dto.VerifyCodeRequest;
+import com.canvi.hama.domain.auth.dto.request.SendCodeRequest;
+import com.canvi.hama.domain.auth.dto.request.VerifyCodeRequest;
 import com.canvi.hama.domain.auth.service.AuthService;
 import com.canvi.hama.domain.auth.service.EmailAuthService;
 import com.canvi.hama.domain.auth.swagger.SendAuthCodeApi;
