+++ src/main/java/com/canvi/hama/domain/auth/dto/response/RefreshTokenResponse.java
@@ -1,4 +1,4 @@
-package com.canvi.hama.domain.auth.dto;
+package com.canvi.hama.domain.auth.dto.response;
 
 public record RefreshTokenResponse(
         String accessToken
