+++ src/main/java/com/canvi/hama/domain/auth/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.canvi.hama.domain.auth.dto;
+package com.canvi.hama.domain.auth.dto.response;
 
 public record TokenResponse(
         String accessToken,
