+++ src/main/java/com/canvi/hama/domain/auth/service/AuthService.java
@@ -4,11 +4,11 @@ import com.canvi.hama.common.exception.BaseException;
 import com.canvi.hama.common.response.BaseResponseStatus;
 import com.canvi.hama.common.security.JwtTokenProvider;
 import com.canvi.hama.common.util.RedisUtil;
-import com.canvi.hama.domain.auth.dto.LoginRequest;
-import com.canvi.hama.domain.auth.dto.LoginResponse;
-import com.canvi.hama.domain.auth.dto.RefreshTokenResponse;
-import com.canvi.hama.domain.auth.dto.ResetPasswordRequest;
-import com.canvi.hama.domain.auth.dto.SignupRequest;
+import com.canvi.hama.domain.auth.dto.request.LoginRequest;
+import com.canvi.hama.domain.auth.dto.response.LoginResponse;
+import com.canvi.hama.domain.auth.dto.response.RefreshTokenResponse;
+import com.canvi.hama.domain.auth.dto.request.ResetPasswordRequest;
+import com.canvi.hama.domain.auth.dto.request.SignupRequest;
 import com.canvi.hama.domain.user.entity.User;
 import com.canvi.hama.domain.user.repository.UserRepository;
 import java.util.Random;
