+++ src/test/java/com/canvi/hama/diary/controller/ImageTest.java
@@ -2,8 +2,8 @@ package com.canvi.hama.diary.controller;
 
 import static io.restassured.RestAssured.given;
 
-import com.canvi.hama.domain.auth.dto.LoginRequest;
-import com.canvi.hama.domain.auth.dto.SignupRequest;
+import com.canvi.hama.domain.auth.dto.request.LoginRequest;
+import com.canvi.hama.domain.auth.dto.request.SignupRequest;
 import com.canvi.hama.domain.auth.service.EmailAuthService;
 import com.canvi.hama.domain.diary.entity.Diary;
 import com.canvi.hama.domain.diary.exception.DiaryException;
