+++ src/test/java/com/canvi/hama/domain/auth/controller/AuthControllerTest.java
@@ -3,9 +3,9 @@ package com.canvi.hama.domain.auth.controller;
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.canvi.hama.domain.auth.dto.LoginRequest;
-import com.canvi.hama.domain.auth.dto.ResetPasswordRequest;
-import com.canvi.hama.domain.auth.dto.SignupRequest;
+import com.canvi.hama.domain.auth.dto.request.LoginRequest;
+import com.canvi.hama.domain.auth.dto.request.ResetPasswordRequest;
+import com.canvi.hama.domain.auth.dto.request.SignupRequest;
 import com.canvi.hama.domain.auth.service.EmailAuthService;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
